epoch halters done
Next step is moving the halters and schedulers outside of run_epochs so they can be passed in as experiment configurations. Then it should be easy to create a LR Range Finder -> OneCyle workflow. At that point I can queue large experiments with many models and not need to tune the LR. I probably need to do WD hyperparam search still, but using the LR Range Finder to evaluate, a grid search might be pretty fast.