Dict arg override for loaded HPs, LR schedule negative fix
When running a small fractional number of epochs that doesn't cleanly divide by minibatch size, LR was often overshooting zero. It is now clamped to zero in those cases. However, the root cause is not addressed. There seems to be an issue with the epoch ending at slightly incorrect time in that same, which was leading to the overshoot.

Also a dict arg to the main function in run.py which allows a dicitonary to override arguments loaded by all other methods.